Try to fix it based on Max's commit
Thank you Max for your commit!

I guess the reason is we use d-flex in button's parent component, which limit the width of button. So what I'm trying to do is to remove d-flex so that block can work, then add a margin to limit the width of its parent component.